requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-openai
pydantic
httpx
google-generativeai
pydantic-settings
textblob
python-dotenv